Adopt gsl::narrow_cast for safer type conversions
Replaced most static_cast<int> and similar casts with gsl::narrow_cast<int> in MapEditorMain.cpp and TileGroup.cc to improve type safety and prevent narrowing errors. Updated CMakeLists.txt to link Microsoft.GSL::GSL for the editor target. Documentation in DEVELOPMENT.md was expanded to describe the use of GSL and the new casting approach, as well as the addition of Expects guards in layer selection helpers.